BackEnd/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==64.0.0
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
oauthlib==3.3.1
openai==2.14.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
python-dotenv==1.2.1
typing_extensions==4.14.1
tzdata==2025.3
urllib3==2.6.2
